requirements.txt
django==2.2.1
djangorestframework==3.9.3
drf-nested-routers==0.91
factory_boy==2.12.0
freezegun==0.3.11